Add Floaterm plugin
Also added keymaps for rust files to run `cargo run` and a global one
for running `lazygit`. This makes Lazygit feel like a vim plugin, but
actually it's an app that has to be installed on the system. I have
installed it on my macbook as well as my windows machine so hopefully
it's a safe requirement.